# docs/swagger.yaml
basePath: /
definitions:
  github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_api_handler.AdminUserProfileResponse:
    properties:
      memberships:
        items:
          $ref: '#/definitions/internal_api_handler.MemberResponse'
        type: array
      payments:
        items:
          $ref: '#/definitions/internal_api_handler.PaymentResponse'
        type: array
      user:
        $ref: '#/definitions/internal_api_handler.AdminUserResponse'
    type: object
  internal_api_handler.AdminUserResponse:
    properties:
      ban_reason:
        type: string
      banned_at:
        type: string
      banned_by:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_api_handler.BulkBanErrorResponse:
    properties:
      message:
        type: string
      user_id:
        type: integer
    type: object
  internal_api_handler.BulkBanResultResponse:
    properties:
      banned:
        items:
          type: integer
        type: array
      failed:
        items:
          $ref: '#/definitions/internal_api_handler.BulkBanErrorResponse'
        type: array
    type: object
...
  internal_api_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_rede_world-cup-quiniela_internal_middleware.ErrorDetail'
    type: object
...
  internal_api_handler.MemberResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      joined_at:
        type: string
      paid:
        type: boolean
      quiniela_id:
        type: integer
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal_api_handler.PageMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
  internal_api_handler.Paged-internal_api_handler_AdminUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_api_handler.AdminUserResponse'
        type: array
      page:
        $ref: '#/definitions/internal_api_handler.PageMeta'
    type: object
...
  internal_api_handler.PaymentResponse:
    properties:
      amount:
        type: integer
      confirmed_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      notes:
        type: string
      quiniela_id:
        type: integer
      reference:
        type: string
      reviewed_by:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  internal_api_handler.banRequest:
    properties:
      reason:
        type: string
    type: object
  internal_api_handler.bulkBanRequest:
    properties:
      reason:
        type: string
      user_ids:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for the World Cup prediction game. Manage fixtures, submit
    score forecasts, and track leaderboards.
  title: World Cup Quiniela API
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      description: Returns a paginated list of all user accounts. Supports optional
      parameters:
      - description: Filter by ban status (true = banned only, false = active only)
        in: query
        name: banned
        type: boolean
      - description: Filter by role (admin, user)
        in: query
        name: role
        type: string
      - description: Search by name or email (partial match)
        in: query
        name: search
        type: string
      - description: Max records per page (default 50, max 200)
        in: query
        name: limit
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.Paged-internal_api_handler_AdminUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - admin-users
  /api/v1/admin/users/{id}:
    get:
      description: 'Returns the full admin view of a user account: base profile, group'
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.AdminUserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - admin-users
  /api/v1/admin/users/{id}/ban:
    delete:
      description: Lifts a suspension from a user account, restoring full access.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.AdminUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Unban a user
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      description: Suspends a user account. Banned users are blocked from all write
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ban reason (required)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.banRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.AdminUserResponse'
        "400":
          description: reason is required
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ban a user
      tags:
      - admin-users
  /api/v1/admin/users/bulk-ban:
    post:
      consumes:
      - application/json
      description: Bans multiple user accounts in a single request. Each ban is
      parameters:
      - description: List of user IDs and ban reason
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.bulkBanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handler.BulkBanResultResponse'
        "207":
          description: Partial success - some bans failed
          schema:
            $ref: '#/definitions/internal_api_handler.BulkBanResultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Caller is not an admin
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "422":
          description: user_ids or reason missing
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk-ban users
      tags:
      - admin-users
...
securityDefinitions:
  BearerAuth:
    description: 'Clerk JWT token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"